Move FileSizeFormatter to a new module for injectable Android utilities.
